requirements.txt
skyfield==1.10
astropy==3.0.3
numpy==1.16.1
pandas==0.20.3
matplotlib==2.2.2
requests==2.12.4
scipy==1.1.0
pyquaternion==0.9.2
setuptools==41.0.1
Pint==0.8.1
PyYAML==5.1.1
